# `spawnWindows`: Improve worst-case performance considerably + tests