Follow up to #13983.

First, the eye-popping benchmark:

```
Benchmark 1: winpathbench-perf.exe
  Time (mean ± σ):     854.3 ms ±   6.0 ms    [User: 116.9 ms, System: 731.9 ms]
  Range (min … max):   845.3 ms … 867.6 ms    10 runs

Benchmark 2: winpathbench-master.exe
  Time (mean ± σ):      8.188 s ±  0.038 s    [User: 0.684 s, System: 7.479 s]
  Range (min … max):    8.137 s …  8.254 s    10 runs

Benchmark 3: winpathbench.bat
  Time (mean ± σ):     851.4 ms ±   4.9 ms    [User: 90.3 ms, System: 741.3 ms]
  Range (min … max):   843.9 ms … 861.1 ms    10 runs

Summary
  'winpathbench.bat' ran
    1.00 ± 0.01 times faster than 'winpathbench-perf.exe'
    9.62 ± 0.07 times faster than 'winpathbench-master.exe'
```

<details>
<summary>Benchmark code</summary>

Compiled with `-lc -OReleaseFast`:

```zig
const std = @import("std");

pub fn main() !void {
    const allocator = std.heap.c_allocator;
    var i: usize = 0;
    const iterations = 100;
    while (i < iterations) : (i += 1) {
        var result = std.ChildProcess.exec(.{
            .allocator = allocator,
            .argv = &[_][]const u8{"hello"},
        }) catch continue;
        defer allocator.free(result.stdout);
        defer allocator.free(result.stderr);

        if (result.term != .Exited or result.term.Exited != 0) {
            return error.CommandFailed;
        }
    }
}
```

The .bat file for comparison:

```bat
FOR /L %%i IN (1,1,100) DO hello
```

</details>

However, this is not representative, and in fact I think it's basically impossible to get a 'representative' benchmark, as the benchmark mostly depends on the number of search paths in the `PATH` env var that must be searched before finding the command, which varies hugely (order of PATH, number of values in PATH, the particular command, etc).

The benchmark above is using a `PATH` with 211 entries (artificially inflated), and it finds the command after searching through 197 of them. Here's a benchmark with only the directory that the command is found in on the `PATH` (so it only checks the cwd and then the PATH entry where it finds it):

```
Benchmark 1: winpathbench-perf.exe
  Time (mean ± σ):     237.9 ms ±   4.1 ms    [User: 48.4 ms, System: 179.9 ms]
  Range (min … max):   231.0 ms … 244.0 ms    11 runs

Benchmark 2: winpathbench-master.exe
  Time (mean ± σ):     271.3 ms ±   1.4 ms    [User: 47.5 ms, System: 195.0 ms]
  Range (min … max):   269.4 ms … 274.1 ms    10 runs

Benchmark 3: winpathbench.bat
  Time (mean ± σ):     278.9 ms ±   1.7 ms    [User: 69.4 ms, System: 179.4 ms]
  Range (min … max):   276.0 ms … 281.1 ms    10 runs

Summary
  'winpathbench-perf.exe' ran
    1.14 ± 0.02 times faster than 'winpathbench-master.exe'
    1.17 ± 0.02 times faster than 'winpathbench.bat'
```

Still slightly faster, but it should demonstrate that this PR greatly improves the worst-case performance but not necessarily the performance of the happy path.

---

That out of the way, here's some details:

The name of the game here is to avoid CreateProcessW calls at all costs, and only ever try calling it when we have a real candidate for execution. Secondarily, we want to minimize the number of syscalls used when checking for each PATHEXT-appended version of the app name.

An overview of the technique used (it was inspired by running [`NtTrace`](https://github.com/rogerorr/NtTrace) on the `winpathbench.bat` file and finding that this is what it was doing):
- Open the search directory for iteration (either cwd or a path from PATH)
- Use NtQueryDirectoryFile with a wildcard filename of `<app name>*` to check if anything that could possibly match either the unappended version of the app name or any of the versions with a PATHEXT value appended exists.
- If the wildcard NtQueryDirectoryFile call found nothing, we can exit early without needing to use PATHEXT at all.

This allows us to use a <open dir, NtQueryDirectoryFile, close dir> sequence for any directory that doesn't contain any possible matches, instead of having to use a separate look up for each individual filename combination (unappended + each PATHEXT appended). For directories where the wildcard *does* match something, we only need to do a maximum of `<number of supported PATHEXT extensions>` more NtQueryDirectoryFile calls.

In addition, we now only evaluate the extensions in PATHEXT that we know we can handle (.COM, .EXE, .BAT, .CMD) and ignore the rest (see https://github.com/ziglang/zig/pull/13983#issuecomment-1356155629 for more details about this).

---

This also adds a standalone test and makes two edge cases match Windows behavior:

- If an invalid executable has the extension .exe and it is attempted to be executed, that is now treated as unrecoverable and InvalidExe is immediately returned no matter where the .exe is (cwd or in the PATH). This matches the behavior of the Windows cmd.exe.
- If the app name contains more than just a filename (e.g. it has path separators), then it is excluded from PATH searching and only does a cwd search. This matches the behavior of Windows cmd.exe.