@squeek502 2022-12-18
An addendum about some tradeoffs:

Out of curiosity, I tried comparing against a Rust version of the benchmark, and I noticed that the Rust version performs considerably better:

```
Benchmark 1: winpathbench-perf.exe
  Time (mean ± σ):     848.9 ms ±   3.6 ms    [User: 136.6 ms, System: 702.8 ms]
  Range (min … max):   841.5 ms … 853.3 ms    10 runs

Benchmark 2: winpathbench-rust.exe
  Time (mean ± σ):     545.4 ms ±   3.8 ms    [User: 81.9 ms, System: 440.3 ms]
  Range (min … max):   539.6 ms … 551.5 ms    10 runs

Summary
  'winpathbench-rust.exe' ran
    1.56 ± 0.01 times faster than 'winpathbench-perf.exe'
```

In looking into why this is, I noticed that Rust makes some tradeoffs in terms of matching Windows behavior and what it thinks is the most likely scenario:

- [When resolving a command without an extension, it will append `.exe` and then only search for that (no other extensions from PATHEXT)](https://github.com/rust-lang/rust/blob/48b3c4612681ba7828880aa9e675452c62714bbf/library/std/src/sys/windows/process.rs#L375-L386). This means that it will only check exactly one file per directory, which I believe is what makes the difference in the benchmark. The tradeoff here is that it will fail to find commands that Windows *would* find, e.g. if there is a `something.bat` in the PATH and you call `something` Rust will not find it.
- Rust also [gets the `Windows\system32` directory and checks it early](https://github.com/rust-lang/rust/blob/48b3c4612681ba7828880aa9e675452c62714bbf/library/std/src/sys/windows/process.rs#L483) regardless of where it is in PATH or if it's there at all. This hugely speeds up finding things when they are in `system32`, but makes it so that Rust would find things in `system32` even if `PATH` is set to an empty string.

So, my conclusion is that performance could be increased if these sorts of tradeoffs are acceptable, but I think I personally prefer matching the Windows behavior even if it comes at a cost.