@matu3ba 2022-12-18
> So, my conclusion is that performance could be increased if these sorts of tradeoffs are acceptable, but I think I personally prefer matching the Windows behavior even if it comes at a cost.

1. I do assume that absolute paths have the same performance?
2. Yes, Rust doing the wrong thing for kernel32 routines is not nice.
3. For relative paths: For a large amount of spawns performance this can be important, for example if one does a pile of panic tests. It sounds okayish, if libstd offers less performance, but not so okay, if creating the code means lots of boilerplate. Or what do you think?

Aside, has very thorough test coverage.