@squeek502 2022-12-18
> I do assume that absolute paths have the same performance?

If the app name is absolute (e.g. `argv = &.{"C:\some\path\something.exe"}`, then it will always do at most `1 + <num of supported PATHEXT extensions` `NtQueryDirectoryFile` calls, and will not search the `PATH` at all. So, absolute paths should always be extremely quick. If the `cwd` is absolute, though, this does not hold (or should not once https://github.com/ziglang/zig/pull/13994 is merged).

> For relative paths: For a large amount of spawns performance this can be important, for example if one does a pile of panic tests. It sounds okayish, if libstd offers less performance, but not so okay, if creating the code means lots of boilerplate. Or what do you think?

Not sure exactly what you're asking, but I'll try to clarify how this PR will impact normal use cases:

- This will significantly speed up the case of a command not being found in the `PATH`. However, this likely won't be too noticeable in real use cases since continuously execing missing commands is not something people try to do very often.
- This will definitely speed up the case of a command being found somewhere near the end of the `PATH` (e.g. if `C:\Windows\system32` is the last entry in the `PATH` and the `PATH` has a lot of entries). This is a much more common/likely scenario and will be the main benefit of this change for normal use.

If you're asking about if there would be a benefit to taking some of the sorts of tradeoffs that Rust makes, they would have some effect, but again (in terms of 'normal' use) probably only in the situations in the two points above. The most noticeable differences are in the outlier cases where there are a ton of entries in the `PATH` that need to be searched, so I'm not convinced that sacrificing correctness (in terms of conforming to how Windows behaves) is worth the speedups in mostly outlier scenarios.